Add First Stub of CV.tex
This repo is inspired by https://github.com/jankapunkt/latexcv. The
design there is beautiful, yet the LaTeX structure is a bit
complicated. Thus, this is an attempt to refactor the code and
use the idea of LaTeX being a markup language. I try to create
environments and commands with logical meaning rather relying overly
on setting spaces here and there by hand.